Puts the simple vocab element filters on Collections as well as Items. See forum conversation at  http://omeka.org/forums/topic/simple-vocab-for-collections
Open question emerging of how complex to make the plugin to handle other content that might make use of Element Sets, or to keep it simple and have the plugin only care about record types in core that use Element Sets, and whether/how to make distinct options for Items vs. Collections.